Update patch set 4
Patch Set 4:

I'm still -2 on that because of all the convoluted logic around boto

we can't be doing that much version shuffle. if the answer is require newer boto, then let's do it in global requirements and be done with it. But deciding to pip install at random versions based on python import errors is the wrong thing

Patch-set: 4
Label: Workflow=0