Update patch set 7
Patch Set 7:

To be honest, I don't think Id point to any of the Neutron plugin systems as 'that's how to do it".  It totally pre-dated everything else and since it is mostly vendor-specific stuff we can't test I've left it alone.  Now you've gone and brought in something that we can not only test but may want in the gate.  These two points make ignoring Ryu totally necessary.

It looks like you look for 'opendaylight' in Q_ML2_PLUGIN_MECHANISM_DRIVERS, could you do this similar to the existing linuxbridge and openvswitch plugins?  It seems like opendaylight is parallel to those, and if some tweakage needs to be make to make that work lets look at it.

I'll be honest, I don't have most of the Neutron bits in DevStack in my brain and it seems to have gotten a little complex.  If these drivers work better in extras.d I think that's a clue...

Patch-set: 7
Label: Code-Review=0
Label: Workflow=0