Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

Brett, the XenServer CI is non-voting but we shouldn't ignore it if there is a real issue.  It looks like the log suffers from the buffering problem we finally fixed yesterday so this comment should re-trigger a new check.  The log stops late in the process so I doubt it is umask-related.

Patch-set: 1
Label: Code-Review=+2