Update patch set 7
Patch Set 7: Restored

So, after spending a lot more time on this, I now am back to thinking extras is the appropriate place for this. The reason is that the OpenDaylight Neutron integration doesn't run anything on compute hosts. We just need to point the Open vSwitch OVSDB and OpenFlow connections at the OpenDaylight controller, and clean them up at unstack time. This patch simply does that.

Patch-set: 7