Adding rsync enablement support for Fedora
The actual way to enable rsync service is check if
is ubuntu, then enable it in /etc/default/rsync
if is not ubuntu, assume that's using xinetd.
However, Fedora is no longer using xinetd, it's
using systemd instead.
This patch check if it's fedora, if so, it uses
systemd to enable the service.

Change-Id: I83fedb91e0e3860c505927c526f8f0909cef5dfe