Update patch set 5
Patch Set 5:

I'm okay with this, but would be nice if we can check if the service is enabled on Fedora (because it's using systemd, and doesn't have /etc/xinetd.d/rsync file, but you can't ensure the service is enabled).
Right now I can enable manually in my image.

Patch-set: 5