Fix logic in is_neutron_enabled
The logic in is_neutron_enabled needs to take into account the case where a
compute host is running with a control node with Neutron, but no Neutron
agents are on the host. This means the multi-node devstack compute nodes will
not have any Neutron code running on them. Modifying "is_neutron_enabled" to
take into account enabling the meta "neutron" service serves this purpose
quite nicely.

Note: This is the case for OpenDaylight. In this case, in the local.conf for
the compute node, all that will be enabled is the "neutron" service, nothing
prefixed with "q-".

Change-Id: If89de49878269b8a185e3a778b2fa42e3da46221