Update patch set 1
Patch Set 1:

Hmm, this didn't actually fix the problem, the jenkins slave must still be too quick.  I'm going to do a test patch that just does a sleep 15 and see if that helps, just to see if this is the problem.  Unfortunately my test system never triggers this failure so I can't debug it there.

Patch-set: 1
Label: Code-Review=0