Update patch set 1
Patch Set 1:

Thanks for the comments dtroyer! So, in looking at is_service_enabled() in functions-common, what I see is that it does first check the base service, in this case neutron, and it sets "enabled" to "0". However, then it calls "is_neutron_enabled, which only checks for "q-" being enabled, which then sets enabled to "1", causing this to fail. I think the fundamental problem is there is no way to tell a compute host it's running Neutron without any neutron services as this is structured. Thus, I think this patch is correct. What do you think?

Patch-set: 1
Label: Code-Review=0