Update patch set 1
Patch Set 1:

For my testing, nova-comptue is all that's running. OpenDaylight uses OpenFlow and OVSDB to talk to each compute host to program networking on the host. libvirt sets the VIF up.

The basic issue here is that in the case of using Neutron with OpenDaylight, there are no Neutron services configured on the host itself. devstack currently makes that assumption. which is thus broken in this case.

Patch-set: 1