Update patch set 3
Patch Set 3:

I changed this back to what was in patch #1, but it still doesn't reliably fix the error.  There's nothing wrong with the code if you still want to apply it, it might just be the firewall_driver setting that's tripping this up?  If that's not set in the gate then no filtering will be done, and the ping *should* succeed.

This also might be related to https://bugs.launchpad.net/neutron/+bug/1210121

Patch-set: 3
Reviewer: Gerrit User 1131 <1131@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0