Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

I'm honestly not all that keen on this, especially coming in on i3 week with an acknowledged core issue in marconi with it incorrectly handling of output. 

This is made more concerning because marconi has insisted on a new wsgi stack (not using one used by other components). So I think it needs to be working correctly, not worked around.

Patch-set: 2
Label: Code-Review=-1