Update patch set 2
Patch Set 2:

re bug landing late, we have been banging on this for weeks and working with others to find the root cause, but the crazy FD redirection devstack does has made this problem extremely difficult to track down. We are filing a bug this late simply because we wanted to first exhaust all ideas.

re WSGI, if you are referring to Falcon, that has absolutely NOTHING to do with this problem and I am frankly tired of these drive-by comments that are anything but constructive.

Patch-set: 2