Update patch set 1
Patch Set 1: Looks good to me (core reviewer); Approved

This doesn't fail in the gate because it's the expression for an if and not considered a failed command.  And because the celio API either always starts in time or doesn't fail in a detectable way.

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1