read returns failure when the docker pid file has no newline
The bash command "read" returns a failure code (1) when it reads EOF
before it reads a newline.  In tests on ubuntu13.10 the docker pid
file does not contain a newline.  This cause the read command to return
a non-zero value and thus stack.sh fails.  This patch replaces the
call to "read" with a call to "cat" in the needed places.

Change-Id: I586d5585db8ad50748ae162e586dc5adeced934d