Update patch set 17
Patch Set 17: I would prefer that you didn't merge this

(8 inline comments)

Does Ironic support more than Ubuntu? If so the prereq files need to be present. If not, there should be some checks to gracefully abort on un-supported platforms.

The scripts in tools/ironic/scripts need to have some semi-reasonable error handling. I'll not hold up this review for that, but please follow up with it soon.

Python scripts need to pass pep8, that check is not built-in to DevStack jobs.

Patch-set: 17
Label: Code-Review=-1
Label: Workflow=0