Update patch set 1
Patch Set 1:

The ryu plugin has exactly two commits since the rename from Quantum to Neutron last July.  The second commit was a recent function header format change.  Maybe there was no need for updates, which of course is good, but that leads me to ask if it still in use since there is no other mechanism to reliably know.

Patch-set: 1