Update patch set 5
Patch Set 5:

We are adding our plugin for the Icehouse release and the plugin code has been merged (https://review.openstack.org/#/c/69246/). We would like to get the install script for the plugin also into the Icehouse release. Please review and approve this change to devstack when you get a chance. Thanks.

Patch-set: 5
Label: Code-Review=0