Update patch set 7
Patch Set 7:

I would like this to be reverted.
We don't have mongodb in power, so this patch is breaking our ci. Sadly we coudln't see this in time because we are not listening all the projects yet.
Unless give us an option to pass even if a package is missing or not.

Patch-set: 7