Make yum install ignore missing packages if architecture is ppc64
There's a few packages missing in ppc64 architecture.
The latest patch https://review.openstack.org/#/c/77067
Broken the devstack in ppc64 due the check if there's
missed packages.
This patch check if the architecture is ppc64 if so
uses the old way, just with yum install -y.
Otherwise, it will use the new check.

Change-Id: I3fd4aeaf183b8def787ba677f8941bf52e5d1b53