Update patch set 2
Patch Set 2:

note: We don't have the grenade-partial-ncpu-forward job running on this to show that this fix definitely works. Furthermore Sean isn't sure that the forward jobs are doing what we expect them to do, so we are taking a bit of risk on this patch.

Patch-set: 2
Reviewer: Gerrit User 1849 <1849@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0