Makes error message easier to understand.
When executing:

$ ./stack.sh

I receive the following error message.

"[Call Trace]
./stack.sh:292:die
[ERROR] ./stack.sh:292 Could not determine host ip address.
Either localrc specified dhcp on or defaulted"

So I tried finding localrc but to no avail.
Then I did a quick search online which produced the url-
http://devstack.org/localrc.html

This url states:

"localrc is the old file used to configure DevStack.
It is deprecated and has been replaced by local.conf.
DevStack will continue to use localrc if it is present and
ignore the localrc section in local.conf.. Remove localrc
to switch to using the new file."

I also did a search in stack.sh and found that the error
message is using ${HOST_IP_IFACE} and on my system this
environment variable was not defined, rendering the error
message even more opaque.

My suggested changes are:

	*Take out the reference to localrc as it is now
	depreciated
	*Point the user towards local.conf where there are
	comments on how to set HOST_IP

Change-Id: I41f14a2de85449d2a08ab7eb2849844a1087b147
Closes-Bug: #1290556