change the vmdk to the one used in VMWare ci
The debian image that defaults to being using with vmware is huge,
and it turns out it's not actually used in VMWare ci so we don't
really know if it's working. Instead use the vmdk that is used
in VMWare ci, which we know will boot, as we get results everyday.

Change-Id: I014746af293852525e2bd128c4d19f5889ecd55d