Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(2 inline comments)

There is at least one bug in the project loop (needs --long to get the enabled status), but I've hated this code for a long time, I've left a couple of suggestions how to clean it up visibly and reduce the number of processes running.

If nothing else, using the OSC output formatting should be useful...there's an example of a ShowOne style (-f shell) and a Lister style (csv) that will clean up all the pipe filters.

Patch-set: 1
Label: Code-Review=-1