Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

IMO This patch should also drop nova-baremetal support entirely in favor of Ironic.  Would probably require an announcement/discussion on the ML, and coordination with infra. if the baremetal stuff is being used for testing anywhere in the gate.

Patch-set: 1
Label: Code-Review=-1