XenAPI: Cirros images must always boot as PV.
The default for VHD disk-types is PV, which is why booting from a
server works.  However, creating a volume from the image needs to
pass this parameter on to the volume.  Note that
Id673158442fde27e8d468ca412c9bd557a886e6b is also required to fix
bug 1294069

Change-Id: I7ea1d85d6082787ac4551f78300a04bf59074261
Partial-Bug: 1294069