Update patch set 3
Patch Set 3:

Is this required on all Ubuntu stack.sh runs?  It adds ~15 min to the runtime, which is not trivial.  If we can find times it is not necessary that would be great.

Also, it maybe should be in lib/nova_plugins/functions-libvirt install_libvirt().

Patch-set: 3
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0