Update patch set 14
Patch Set 14: Looks good to me, but someone else must approve

I have a similar line of questioning as Sebastien since I proposed patches back in August for similar purposes that weren't even given the courtesy of a response.

Since then I've been contacted by several companies working on openstack that wish to integrate ceph offerings and have shared my patchsets with them. We also use a patchset on top of devstack at Metacloud to develop our ceph integration with Nova.

If devstack is here to help push forward the core development of openstack, it seems counterproductive to resist supporting a feature that is currently in desperate need of more core development. 

Keeping this code in a separate repo seems like it will send all of the users on a duck hunt to look at one or more external repos to get a working stack, also it encourages forking and creating multiple overlapping patchsets to accomplish the same goal as we have today.

Patch-set: 14
Label: Code-Review=+1