Update patch set 14
Patch Set 14: I would prefer that you didn't merge this

Sébastien, David,

Every component that gets added to DevStack must be maintained.  And one of the filters we use to decide what we add is a dependency on external repositories.  We have limited resources (who doesn't?) and have to prioritize what gets done.

This is the primary reason for creating the extras.d plugins, so additional features can be dropped in easily without requiring people to maintain DevStack forks, just a normal repo or tarball of the additional bits.  It isn't perfect, but I'm more than happy to look at the things that might be needed to make it work in new circumstances.

It looks the changes I asked for have not been made yet, specifically using the plugin model that the Nova hypervisor plugins have rather than the database model which I think is unnecessarily complicated.  If my mind needs to be changed I'll consider specific reasons in favor of leaving it as is.

Personally, I really want to try Ceph under DevStack and have not made the time to do it yet.

David, I apologize if you feel slighted regarding your review in August.  Many times proposals for new features and whatnot slide through the cracks in the volume when we are not expecting them or have even heard the use case for the feature.  By way of example, I worked with Sam Alba for some time before he got the Docker bits in.

Patch-set: 14
Label: Code-Review=-1
Label: Workflow=0