Update patch set 3
Patch Set 3:

I have to confirm that the necessary change is in the client to deal with both v2 and v3  if the endpoint says v2.0.  Is so, the keystone piece on this change probably is not necessary.

Patch-set: 3
Reviewer: Gerrit User 2218 <2218@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0