Make keystone.conf world readable
The keystone.conf file was set to 0600 and only the user could
read it. The user that Tempest runs as could not read it and a test
for `keystone-manage db_version` would fail because it can't read
the config file.

The keystone.conf file will now be 644 so that the Tempest user
can read it and we can add tests for keystone-manage.

The keystone.conf file should now match the nova.conf file
permissions. There are already tests for nova-manage in Tempest.

Change-Id: I4ce406f32189a81c6a7361a11dc474886629b806
Closes-Bug: #1284341