Provide a keystone.conf for tempest
The keystone.conf file isn't readable by the tempest user so it
can't run commands like `keystone-manage db_version`. This is for
security reasons.

This change adds a keystone.tempest.conf file that tempest has
access to. It only has the database connection string (so it
doesn't include the admin token or ldap password).

Closes-Bug: #1284341
Change-Id: I4ce406f32189a81c6a7361a11dc474886629b806