Provide a keystone.conf for tempest
The keystone.conf file isn't readable by the tempest user so it
can't run commands like `keystone-manage db_version`. This is for
security reasons.

This change adds a keystone.tempest.conf file that tempest has
access to. It only has the database connection string (so it
doesn't include the admin token or ldap password).

Change-Id: I5180be85fd6449c16f531580a1ddd1d11bfd2572
Closes-Bug: #1284341