Update patch set 9
Patch Set 9: I would prefer that you didn't merge this

in talking with @jgriffith today, it looks like this will be the default in juno. So I'd rather just wait until juno and remove all the tgt code and default to lioadm at that point.

Patch-set: 9
Label: Code-Review=-1
Label: Workflow=0