Update patch set 9
Patch Set 9: I would prefer that you didn't merge this

I'd prefer we avoid adding this like this until (if) we actually change the default target to LIO.  For now this is just a convenience factor for the few that run/test with it locally so there's no reason to not have your own custom patch or fork even to facilitate that.

I think we'll take a look at making this happen in Juno and should flip the switch then.

Patch-set: 9
Label: Code-Review=-1