Update patch set 4
Patch Set 4:

If you go ahead with this I'd like to see 'OPENSTACKAUTH' changed to be more specific as it is fairly generic and a Horizon component is not the first thing that comes to my mind with that name.  Something a bit more specific such as 'DJANGOAUTH' or 'HORIZONAUTH' or 'HORIZON_DJANGO_UNCHAINED_OPENSTACK_AUTH' would work for me.

(That last one is just a joke...)

Patch-set: 4
Label: Code-Review=0