Update patch set 14
Patch Set 14:

Dean, I think rejecting this is a mistake. You talk about maintainability and specifically mention the external repositories.

First of all, there's prior art. The docker backend also adds an external repository and anynoe using devstack on Debian uses external repositories to do so.

Second of all, rejecting it on grounds of maintainability implies that you don't believe Sébastien will stick around to maintain this. In general, I think we should give people the benefit of the doubt (how else will they prove that they'll live up to the responsibility?) and deal with it if the contributor turns out to not actually maintain their code over time (like we did with the Hyper-V driver in Nova, for instance).

For someone like Sébastien, who has demonstrated a consistent interest in and commitment to making the Ceph/OpenStack experience excellent, it particularly strange to question whether he'll ensure this is well maintained. Should he get hit by a bus, though, I and a number of other people on my team are also very interested in getting this included and would be happy to help maintain it.


In short: I think the only fair thing to do here is to review the change based only on its technical quality and assume that this isn't just a fire-and-forget contribution. If you have technical concerns, let's raise them and see if we can work on addressing them.

Patch-set: 14
Label: Code-Review=0