Devstack support for Neutron VPNaaS
Includes three enhancements to support VPNaaS and allow other
vendor implementation support.

First, the /opt/stack/neutron/etc/vpn_agent.ini file is copied
to /etc/neutron, to allow the selected device drivers for agent
to be specified. Both reference and vendor implementations can
be uncommented in this file, prior to stacking.

Second, neutron.conf is modified to provide the VPN service
drivers that are available. The reference implementation
will be selected as the "default" service driver. In the future,
if the service type framework (or flavor based mechanism) is
adopted, any additional service driver implementations can be
specified and will be available. For now though, if an alterative
service driver is desired, it would need to be added and marked
with a ":default" suffix.

Third, to allow vendor VPN implementations to provide an .ini
file to the agent at start-up time, an environment variable is
defined and can be set in the localrc to specify one or more
.ini files. The desire is NOT to include vendor settings in the
global vpn_agent.ini.

Change-Id: I5370ee57bbe326eff505a53055da4fddc0838de4
Closes-Bug: 1301552