Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

I feel like this should be done entirely as an extras.d file which is actually in the docker driver tree. If there are call points which are needed in extras.d, we should add those.

Patch-set: 3
Label: Code-Review=-1
Label: Workflow=0