Update patch set 16
Patch Set 16: Looks good to me, but someone else must approve

"In short: I think the only fair thing to do here is to review the change based only on its technical quality and assume that this isn't just a fire-and-forget contribution. If you have technical concerns, let's raise them and see if we can work on addressing them."

Soren's comment really summarizes the core of this back-and-forth quite well.  Reviews such as this should be an objective process, with acceptance or rejection based on the quality of the implementation and likely use of the functionality in question.  There are a number of developers working on Ceph-based functionality and having a common framework to use for development will serve to improve the quality of the integration across the board.

I would like to echo Soren's view that rejecting this submission based on unfounded concerns over maintenance  abandonment would be a mistake.  We should be erring towards flexibility and ripping out unmaintained/broken functionality, rather than rejecting it over the possibility that someday it might reach that state.  If nothing else, the fact that this review has been going back and forth for months, with a willingness to resolve technical issues,  should prove that it's not going to be abandoned (not that I think the counter assumption is the slightest bit fair).

Patch-set: 16
Reviewer: Gerrit User 1525 <1525@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1