Update patch set 16
Patch Set 16:

My apologies for casting doubt on anyones ability or desire to maintain code, that was certainly not my intent.  And reading that as the reason for not approving this review is simply wrong;  Sean's comment on Jan 11 is still the primary reason.  I have asked more than once for this to be implemented as a drop-in feature and that has still not been done.  This sort of need is exactly the reason for implementing that plugin mechanism.

* The appropriateness of including support for Ceph in DevStack is not a settled question.  Ceph is not an OpenStack project, integrated, incubated or otherwise.  And as far as I know, there is no intention of it becoming one.  You are more than welcome to gauge the desires of the community via the openstack-dev mailing list.
* You will note that ALL of the other non-OpenStack projects DevStack supports/depends (now that Docker is gone) are available in the common distribution repositories.  We use that as one filter to reduce the number of variables that we support directly in the DevStack repo.
* I have not been approached by ANYONE until Sebastien's comments on Mar 21 regarding reasoning or 'what would it take' sorts of questions about this review or Ceph support in general.
* Docker has been removed from DevStack as it has been removed from Nova.  It still required a third-party repo and had not been addressed as was promised to me when we merged the original code.

Patch-set: 16
Label: Code-Review=0
Label: Workflow=0