Update patch set 16
Patch Set 16:

I understand that Ceph is not an OpenStack project and maybe never will be. The same is true for iscsitarget and lvm (the default implementations that Ceph replaces). Ceph is has been included in Fedora and Ubuntu for years already. Do you have more details on the specific criteria that must be met here? Specifically which distros and which releases must include a specific piece of software before it's ok to install it from Devstack?

I have to wonder, though, why such apparently rigid criteria are applied to e.g. Ceph, but we're perfectly content pulling in dependencies from Pypi that are just a few hours old if it fills a need.

Back when we were just starting OpenStack, we chose to target whatever Ubuntu version was most recently released. That gave us a very up-to-date selection of software that we could use. This allowed us to move fast, because we didn't have to worry about older distros not having some dependency that we needed.  With most of our dependencies being pulled from Pypi now, we've actually gone even further in the don't-care-if-some-version-of-some-distro-doesn't-have-this-dependency direction for *Python* dependencies, so it seems remarkable that we're apparently going the exact opposite direction for non-Python stuff.

We rely on recent version of libvirt, qemu, openvswitch, etc., etc.. What's different about Ceph?

When you suggest gauging the desires of the community on the mailing lists, are you referring to the desire of making Ceph an OpenStack project or the desire to have Ceph support in devstack? I'd assume the latter, but given the context where you wrote it, I'm not sure.

Patch-set: 16
Reviewer: Gerrit User 518 <518@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0