Update patch set 16
Patch Set 16:

@Seb What about not using ceph repository but distros specific repository?

@Sean, Dean

The way I think about ceph support here is not so much different than the others drivers we support for cinder/neutron or even for backends like postgresql/mysql. 

Ceph is definitively something popular out there and having devstack support out of the box will help users see how it works with OpenStack.

On the other side I understand that we don't want to have *anything* and *everything* in devstack and just let folks plugin into our plugins system but my guts feeling is that ceph is too much popular to be anything.

(and FWIW I am willing to take it as 'reviews sponsor' for ceph if that's needed.)

Patch-set: 16
Label: Code-Review=0
Label: Workflow=0