Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

Better to split this into two changes as discussed in IRC: one to add the new download (so that we cache it on nodepool images), the other to use it in devstack. That way we can merge them with enough time in between to ensure (automatically or manually) that nodepool devstack images are updated successfully in all providers.

Patch-set: 2
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1