Update patch set 5
Patch Set 5:

At some point in time, this was needed to work around an issue wherein running devstack with the local virbr that Ironic uses, within a Precise node on HP Cloud, would cause sshd to jump onto the bridge and lose its binding to the public IP, rendering the node unreachable.

If that's been addressed elsewhere, then removing this whole section and using the default :22 should be fine.

Patch-set: 5