Fix is_service_enabled backward compat for swift
When localrc have swift but not a s-* in ENABLED_SERVICES,
is_service_enabled returns 1 instead of 0.  is_swift_enabled only checks
for s-* services and return 1 in this case (this the new desired behavior)
but is_service_enabled fails to handle this backward compatibily

This patch add backward compatibility, so when user have only set swift in ENABLED_SERVICES
and a devstack script use 'is_service_enabled swift' it returns 0 now.

Change-Id: I045037004236fc7d9ebb0e71923433f05fdc9a80
Closes-bug: #1279384