Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

(1 inline comment)

@Sergey, the current behavior has the potential to break things -- this patch is one way to correct it, but not the only way.

As I stated previously, the background for doing any of this in the first place is that, in our early testing before this code landed in devstack, if sshd_config did not have any explicit ListenAddress directive, as it does not in HPCS Precise instances, then when starting devstack, sshd would stop listening on eth0 and we'd lose connectivity to the host entirely. We worked around this by explicitly adding eth0's IP as a ListenAddress in sshd_config.

But, folks raised a potential problem with this approach, and this patch is trying to fix that problem now.

Option #1:

If sshd_config does not already contain an explict ListenAddress, then because the current lib/ironic is adding one, it may become the *only* one. This patch proposes to explicitly add the default back.

Option #2:

Another approach would be to remove that section, so that lib/ironic does not add anything to sshd_config. We would need to update some of the tooling to enroll nodes with ssh_port=22. But, this is likely to re-expose the bug we encountered in HPCS Precise instances, causing a total loss of external ssh access to the nodes (eg, locking out jenkins).


So, which approach do you prefer? Explicitly listening on all IPs, or implicitly doing that until something changes? I prefer Option #1 because it is explicit (and that's what this patch is proposing).

Patch-set: 5
Reviewer: Gerrit User 2889 <2889@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1