Add support of hbase database
We have done wrapper and lib for installing nosql databases
(hbase for now) they can be executed via enable_serivce hbase
in localrc.

HBase and Cassandra are supported as database backends in
Ceilometer, MagnetoDB projects, and it can be useful for
development and gating to have it in devstack.

We have implemented Cassandra as part of MagnetoDB devstack integration (
https://github.com/stackforge/magnetodb/tree/master/contrib/devstack) and
started working on HBase now

Here are several blueprints on this occasion
- MagnetoDB
  https://blueprints.launchpad.net/magnetodb/+spec/database-plugin-hbase
  https://blueprints.launchpad.net/magnetodb/+spec/devstack-add-hbase
- Ceilometer
  https://blueprints.launchpad.net/ceilometer/+spec/hbase-tests-on-real-backend

Implements: blueprint add-hbase-to-devstack
Co-Authored-By: Aleksey Chuprin <aleksei_chuprin@symantec.com>
Change-Id: I03fe2ce665b8b21371ee171c1b9f760534c3f2a9