Update patch set 1
Patch Set 1:

Hi Sean, Mehdi's right,

Currently, swift services are launched with "swift" as part of enabled services, and this is broken.

If this feature is not supported, it should fail at start as such with an explicit message and should not start swift services at all to end up miss-configuring them and fail without an explanation.

The best would be to currently use Mehdi's fix and to display a DEPRECATION warning mentioning it will be abandoned in the future, and what will replace it.

Then a future update would totally abandon "swift" as part of enabled services and therefore not try to start them (and miss-configure them).

Patch-set: 1