Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I understand that backporting a new service is rather weird but I am +1 on this change.

The reason, as stated by Jakub, is to ensure the neutron grenade job will succeed. It is currently failing because the db migrations expect metering database table to be present whereas they aren't, since they are created in the havana migration path.

I understand that the best solution to this problem is not to add another service to devstack, but probably to rework neutron's migration path in order to have per-service paths rather than a global one; however, that is a non-trivial solution.

Stability wise, the metering service plugin went through little changes in the icsehouse release cycle, so its reliability in stable/havana is expected to have similar levels to icehouse.

I would therefore suggest to backport the metering server to devstack for stable/havana to the aim of making the neutron grenade job voting in a short time.

Patch-set: 1
Reviewer: Gerrit User 261 <261@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1