Update patch set 1
Patch Set 1:

I think we should just drop the 'swift' service and just make people use the s-* it may be a pain for people to migrate config file but that's  expected for a moving target like devstack. the 'swift' service was only there to easy the migration when we introduced the s-* split but not to be supported as is forever.

Patch-set: 1
Label: Code-Review=0