Update patch set 5
Patch Set 5:

Ah good.  I did see three other failures with this signature on different changesets, but they haven't been occurring recently.  There was also a period of downtime for the CI over last weekend, so a combination of these might have let some bug creep in that was reproducible for XS but only a race for KVM.
That's my theory!

Patch-set: 5