Add IPv6 support
A couple of things are covered in this patch
   -- the openstack management network can be either ipv4 or ipv6
   -- the tenant data network can be either ipv4 or ipv6

Define MGMT_NET with one of the two values 4 or 6 in your localrc
to indicate if you intend to run your management network as either
ipv4 or ipv6. Default value is 4

It's recommended to use host names to define HOST_IP, SERVICE_IP,
etc., in your localrc instead of IPv6 addresses. Those host names
should correspond to ipv6 addresses.

Define IP_VERSION with one of the two values 4 or 6 in your localrc
to indicate if you intend to run your tenant data network as either
ipv4 or ipv6. Default value is 4

Define WITH_IPV6_PATCH with one of the two values 'True' or 'False'
in your localrc to indicate if you inten to run your openstack with
IPv6 patch or not. Default value is False. If the value is True,
then make sure your repository have the following patches enabled:
  - https://review.openstack.org/#/c/52983: may be disabled. So
    make sure it's enabled in your repo
  - https://review.openstack.org/#/c/70649/: may not be working
    without some change.

A couple of variables should be defined in your localrc:
  - IPV6_PRIVATE_RANGE: the ipv6 prefix for your tenant network
  - IPV6_NETWORK_GATEWAY: the gateway IP with the same prefix

If you run without the IPv6 patch, you may want to run a radvd
service in your router namespace to advertise default routes
into your VMs. Define a /etc/radvd.conf such as it contains:
    interface qr-xxxxxxxx-xx
    {
        AdvSendAdvert on;
        MinRtrAdvInterval 3;
        MaxRtrAdvInterval 10;
    };

And then run the radvd inside your router namespace as
   sudo ip netns exec <your router namespace> service radvd start

If you run with the IPv6 patch, you may define two additional
variables in your localrc:
  - IPV6_RA_MODE (with default as slaac)
  - IPV6_ADDRESS_MODE (with default as slaac)

Checkout this file
https://www.dropbox.com/s/9bojvv9vywsz8sd/IPv6%20Two%20Modes%20v3.0.pdf
for the possbile combinations of the two values

It's recommended to run the lastest version of dnsmasq 2.68.

If you intend to enable internet access in your VM, make sure
your network node has ipv6 internet access, and the ipv6 prefix for
your tenant network is a GUA and routable.

Implements: blueprint ipv6-support
Change-Id: I848abf18e00e2a869697c5ef6366bc567dde448a