Let lib/tempest restore errexit setting
If errexit is on, lib/tempest currently fails to restore it after
turning it off.

lib/tempest uses this code to save the current setting for errexit and
restore it later on:

errexit=$(set +o | grep errexit)

This mechanism works works for other settings, such as xtrace, but for
errexit it returns "set +o errexit" even if errexit is on.

Confirmed for:
bash 4.2.25, Ubuntu 12.04.4 LTS (precise)
bash 4.2.37, Debian 7.4 (wheezy)
bash 4.2.45, Fedora 20 (Heisenbug)

Change-Id: Ia477e04567046a7fcdedf699423165bc577b72ad