Let lib/tempest restore errexit setting
If errexit is on, lib/tempest currently fails to restore it after
turning it off.

If tempest is enabled, this affects code running after 80-tempest.sh in
"Phase: extra", "Local Configuration", and "Run local script" (after
that, "Fin" disables errexit unconditionally).

lib/tempest uses this code to save the current setting for errexit and
restore it later on:

errexit=$(set +o | grep errexit)

This mechanism works works for other settings, such as xtrace, but for
errexit it returns "set +o errexit" even if errexit is on.

Change-Id: Ia477e04567046a7fcdedf699423165bc577b72ad