Set allow_migrate_to_same_host as True by default
After OpenStack was installed with devstack on a single node, cold
migration always failed because the default value of
allow_migrate_to_same_host is False, this means that devstack do not
support cold migrate a VM to a same host by default.

Normally, a developer often set up OpenStack with devstack on a single
node, so it is better that we change this value to True to enable
cold migration with a single node after installed by devstack.

Change-Id: I961395f1c79d9600917b6186964c13c92a8a6266