Update patch set 1
Patch Set 1:

Yeah, sorry the commit message is not clear. Let me try to explain.

Devstack Readme (here https://github.com/openstack-dev/devstack#swift-s3) says that to enable the swift3 middleware (swift3 middleware is just a middleware to swift, it's not a binary nor a "service") one must have "enable_service swift3" in his localrc.

But commit I78222d7ee2ea8a281962129d5bcf06c118469510 prevents the presence of any string led by the prefix "swift" in the ENABLED_SERVICES variable. (because of the usage of the shell operator =~).

As "swift" is a prefix of "swift3", it's not possible to install swift3 anymore. Thus this patch (or a variant) is required.

Patch-set: 1