Also pass FWAAS conf to VPN agent if enabled
If q-vpn is enabled, the Neutron neutron-vpn-agent is started
instead of the neutron-l3-agent, but only neutron-l3-agent receives
the fwaas_driver config when q-fwaas is enabled.  This ensures the FW driver
config is passed to either so that a properly configured agent is running
when the FWAAS extensions are enabled.

Change-Id: I237d2831a6b87ae46dfef047d3dc0b85fa6423d3