Update patch set 6
Patch Set 6: I would prefer that you didn't merge this

I do not like the model of plugin that lib/databases implements and wish it would be changed so other projects do no continue to copy it as a starting point.

This also requires another third-party repo for installation.

Please implement it as an extras.d plugin, I do not see anything here that can not be handled that way.  At that point it will be a purely drop-in addition to DevStack.  Inclusion in the repo is not assured, per current practice.  FWIW, there will be a session at the Summit to refine that policy.

Patch-set: 6
Label: Code-Review=-1
Label: Workflow=0