Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Jordan, yes, the problem with that patch is that the 'swift' string is not completely anchored in the comparison.  Changing line 153 in https://review.openstack.org/#/c/87140/1/stack.sh,unified to be:

 if [[ ,${ENABLED_SERVICES}, =~ ,"swift", ]]; then

(note the additional commas) should take care of the problem.  'swift3' should be a valid "service" name.

Patch-set: 1
Label: Code-Review=-1