Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(4 inline comments)

Ugh I just noticed this is against stable/havana.  Everything below was written assuming it was for master, I don't see this getting merged into an old stable branch at all...but it should drop-in just fine.
-----

Overall, this looks good.  Since you are installing from distro packages (yay!) please add files/*/sheepdog prereq files listing the packages required (just sheepdog it appears?).  This helps get the packages in place in the CI setup.  It should look like:

  sheepdog  # NOPRIME

The 'NOPRIME' is a signal to not pre-install the package as it starts a service and there are times we don't want that.

I am still not convinced that extras like this belong in the DevStack repo, but if they are going to get in, this is how they should look.  Thanks.

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0